refactor(workouts): change workouts to workouts.schema
