+++ src/lib/db/index.ts
@@ -5,7 +5,7 @@ import * as AuthSchema from "./schema/auth"
 import * as MuscleSchema from "./schema/muscles"
 import * as ExerciseSchema from "./schema/exercises/schema"
 import * as ExerciseMusclesSchema from "./schema/exercises/muscleGroups"
-import * as WorkoutSchema from "./schema/workouts"
+import * as WorkoutSchema from "./schema/workouts/schema"
 
 const sqlite = new Database(process.env.DB_URL!)
 export const db = drizzle(sqlite, {
