+++ src/lib/db/schema/auth.ts
@@ -7,7 +7,7 @@ import {
 import { createInsertSchema, createSelectSchema } from "drizzle-zod"
 import type { AdapterAccount } from "@auth/core/adapters"
 import { relations } from "drizzle-orm"
-import { workouts } from "./workouts"
+import { workouts } from "./workouts/workout.schema"
 
 export const users = sqliteTable("user", {
   id: text("id").notNull().primaryKey(),
