+++ src/lib/repositories/getWorkouts.ts
@@ -2,7 +2,7 @@ import { eq } from "drizzle-orm"
 
 import { db } from "@/lib/db"
 import { SelectUser } from "@/lib/db/schema/auth"
-import { SelectWorkout, workouts } from "@/lib/db/schema/workouts"
+import { SelectWorkout, workouts } from "@/lib/db/schema/workouts/workout.schema"
 
 export async function getWorkoutsByUserId(
   userId: typeof SelectUser._type["id"]
